requirements.txt
fastapi
uvicorn
opencv-python-headless
Pillow
torch
torchvision
ultralytics
python-multipart
